requirements.txt
awesome-slugify==1.6.5
Django==1.10.6
django-bootstrap3==8.2.3
django-extensions==1.7.9
django-jquery==3.1.0
et-xmlfile==1.0.1
jdcal==1.3
mysqlclient==1.3.10
olefile==0.44
openpyxl==2.4.8
Pillow==4.0.0
pytz==2016.10
regex==2017.6.7
six==1.10.0
Unidecode==0.4.20
psycopg2==2.7.3.2
gunicorn>=19.7.1
django-crispy-forms
